gegl_color_get_rgba() takes pointers to doubles now, not floats.
2009-01-25  Michael Natterer  <mitch@gimp.org>

	* app/tools/gimpgegltool.c (gimp_param_spec_duplicate):
	gegl_color_get_rgba() takes pointers to doubles now, not floats.


svn path=/trunk/; revision=27944

+++ app/tools/gimpgegltool.c
@@ -546,10 +546,10 @@ gimp_param_spec_duplicate (GParamSpec *pspec)
     {
       GeglColor *gegl_color;
       GimpRGB    gimp_color;
-      gfloat     r = 0.0;
-      gfloat     g = 0.0;
-      gfloat     b = 0.0;
-      gfloat     a = 1.0;
+      gdouble    r = 0.0;
+      gdouble    g = 0.0;
+      gdouble    b = 0.0;
+      gdouble    a = 1.0;
       GValue     value = { 0, };
 
       g_value_init (&value, GEGL_TYPE_COLOR);
